refactor(codec): own determinism primitives
