+++ crates/copybook-codec/Cargo.toml
@@ -29,7 +29,6 @@ copybook-overpunch.workspace = true
 copybook-fixed.workspace = true
 copybook-rdw.workspace = true
 copybook-corruption.workspace = true
-copybook-determinism.workspace = true
 copybook-error.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -37,6 +36,7 @@ thiserror.workspace = true
 tracing.workspace = true
 base64.workspace = true
 sha2.workspace = true
+blake3.workspace = true
 copybook-codec-memory.workspace = true
 indexmap.workspace = true
 metrics = { workspace = true, optional = true }
@@ -49,6 +49,7 @@ anyhow.workspace = true
 assert_cmd.workspace = true
 predicates.workspace = true
 copybook-options = { path = "../copybook-options" }
+copybook-determinism = { path = "../copybook-determinism" }
 
 [features]
 default = ["comp3_fast"]
