+++ crates/copybook-codec/src/determinism.rs
@@ -8,10 +8,160 @@
 use crate::lib_api::{decode_record, encode_record};
 use crate::options::{DecodeOptions, EncodeOptions};
 use copybook_core::{Error, ErrorCode, Result, Schema};
-use copybook_determinism::compare_outputs;
 use copybook_rdw::RdwHeader;
 
-pub use copybook_determinism::{ByteDiff, DeterminismMode, DeterminismResult};
+/// Default cap used when collecting byte-level differences.
+pub const DEFAULT_MAX_DIFFS: usize = 100;
+
+/// Hex-encoded BLAKE3 digest length in characters.
+pub const BLAKE3_HEX_LEN: usize = 64;
+
+/// Mode of determinism checking (decode-only, encode-only, or full round-trip).
+#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum DeterminismMode {
+    /// Check that decoding the same binary data twice produces identical JSON.
+    DecodeOnly,
+    /// Check that encoding the same JSON twice produces identical binary data.
+    EncodeOnly,
+    /// Check that decode→encode→decode produces identical JSON.
+    RoundTrip,
+}
+
+/// Details about a byte difference found during determinism checking.
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct ByteDiff {
+    /// Byte offset where the difference was found.
+    pub offset: usize,
+    /// Byte value from the first run.
+    pub round1_byte: u8,
+    /// Byte value from the second run.
+    pub round2_byte: u8,
+}
+
+/// Result of a determinism check operation.
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct DeterminismResult {
+    /// The mode of checking that was performed.
+    pub mode: DeterminismMode,
+    /// BLAKE3 hash of the first run's output.
+    pub round1_hash: String,
+    /// BLAKE3 hash of the second run's output.
+    pub round2_hash: String,
+    /// Whether the two runs produced identical outputs.
+    pub is_deterministic: bool,
+    /// If non-deterministic, details of the byte differences.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub byte_differences: Option<Vec<ByteDiff>>,
+}
+
+impl DeterminismResult {
+    /// Returns true if both runs produced identical outputs.
+    #[must_use]
+    #[inline]
+    pub fn passed(&self) -> bool {
+        self.is_deterministic
+    }
+
+    /// Returns the number of byte differences found (0 if deterministic).
+    #[must_use]
+    #[inline]
+    pub fn diff_count(&self) -> usize {
+        self.byte_differences.as_ref().map_or(0, Vec::len)
+    }
+}
+
+/// Compute a lowercase hex BLAKE3 hash for a byte slice.
+#[must_use]
+#[inline]
+pub fn blake3_hex(data: &[u8]) -> String {
+    blake3::hash(data).to_hex().to_string()
+}
+
+/// Compare two byte slices and build a determinism result with the default diff limit.
+#[must_use]
+#[inline]
+pub fn compare_outputs(mode: DeterminismMode, round1: &[u8], round2: &[u8]) -> DeterminismResult {
+    compare_outputs_with_limit(mode, round1, round2, DEFAULT_MAX_DIFFS)
+}
+
+/// Compare two byte slices and build a determinism result with an explicit diff limit.
+#[must_use]
+pub fn compare_outputs_with_limit(
+    mode: DeterminismMode,
+    round1: &[u8],
+    round2: &[u8],
+    max_diffs: usize,
+) -> DeterminismResult {
+    let hash1 = blake3::hash(round1);
+    let hash2 = blake3::hash(round2);
+    let is_deterministic = hash1 == hash2;
+
+    DeterminismResult {
+        mode,
+        round1_hash: hash1.to_hex().to_string(),
+        round2_hash: hash2.to_hex().to_string(),
+        is_deterministic,
+        byte_differences: if is_deterministic {
+            None
+        } else {
+            Some(find_byte_differences_with_limit(round1, round2, max_diffs))
+        },
+    }
+}
+
+/// Find byte-level differences between two slices using [`DEFAULT_MAX_DIFFS`] entries at most.
+#[must_use]
+#[inline]
+pub fn find_byte_differences(round1: &[u8], round2: &[u8]) -> Vec<ByteDiff> {
+    find_byte_differences_with_limit(round1, round2, DEFAULT_MAX_DIFFS)
+}
+
+/// Find byte-level differences between two slices with an explicit limit.
+#[must_use]
+pub fn find_byte_differences_with_limit(
+    round1: &[u8],
+    round2: &[u8],
+    max_diffs: usize,
+) -> Vec<ByteDiff> {
+    if max_diffs == 0 {
+        return Vec::new();
+    }
+
+    let min_len = round1.len().min(round2.len());
+    let max_len = round1.len().max(round2.len());
+    let mut diffs = Vec::with_capacity(max_diffs.min(max_len));
+
+    for (offset, (&byte_a, &byte_b)) in round1.iter().zip(round2.iter()).enumerate() {
+        if byte_a != byte_b {
+            diffs.push(ByteDiff {
+                offset,
+                round1_byte: byte_a,
+                round2_byte: byte_b,
+            });
+            if diffs.len() >= max_diffs {
+                return diffs;
+            }
+        }
+    }
+
+    if round1.len() != round2.len() {
+        for offset in min_len..max_len {
+            let byte_a = round1.get(offset).copied().unwrap_or(0);
+            let byte_b = round2.get(offset).copied().unwrap_or(0);
+            diffs.push(ByteDiff {
+                offset,
+                round1_byte: byte_a,
+                round2_byte: byte_b,
+            });
+            if diffs.len() >= max_diffs {
+                return diffs;
+            }
+        }
+    }
+
+    diffs
+}
 
 fn serialize_json(value: &serde_json::Value, context: &str) -> Result<Vec<u8>> {
     serde_json::to_vec(value).map_err(|e| {
@@ -252,6 +402,34 @@ mod tests {
         let result = compare_outputs(DeterminismMode::DecodeOnly, &bytes1, &bytes2);
         assert!(!result.is_deterministic);
         assert!(result.diff_count() > 0);
+    }
+
+    #[test]
+    fn primitive_comparison_reports_bounded_differences() {
+        let result =
+            compare_outputs_with_limit(DeterminismMode::EncodeOnly, b"ABCDEF", b"ABxDEy", 1);
+
+        assert!(!result.passed());
+        assert_eq!(result.diff_count(), 1);
+        assert_eq!(
+            result.byte_differences.as_ref().expect("diffs")[0].offset,
+            2
+        );
+    }
+
+    #[test]
+    fn primitive_hash_and_result_serde_are_stable() {
+        let hash = blake3_hex(b"copybook");
+        let result = compare_outputs(DeterminismMode::RoundTrip, b"copybook", b"copybook");
+
+        assert_eq!(hash.len(), BLAKE3_HEX_LEN);
+        assert_eq!(result.round1_hash, hash);
+        assert!(result.byte_differences.is_none());
+
+        let json = serde_json::to_string(&result).expect("serialize determinism result");
+        let decoded: DeterminismResult =
+            serde_json::from_str(&json).expect("deserialize determinism result");
+        assert_eq!(decoded, result);
     }
 
     #[test]
