+++ tests/proptest/Cargo.toml
@@ -32,7 +32,6 @@ copybook-safe-index.workspace = true
 copybook-safe-text.workspace = true
 copybook-corruption.workspace = true
 copybook-corruption-rdw.workspace = true
-copybook-determinism.workspace = true
 copybook-rdw-predicates.workspace = true
 copybook-contracts.workspace = true
 copybook-governance-grid.workspace = true
