+++ tests/proptest/prop_determinism.rs
@@ -9,11 +9,11 @@
 
 #![allow(clippy::expect_used, clippy::unwrap_used)]
 
+use copybook_codec::determinism::blake3_hex;
 use copybook_codec::{
     Codepage, DecodeOptions, EncodeOptions, RecordFormat, decode_record, encode_record,
 };
 use copybook_core::parse_copybook;
-use copybook_determinism::blake3_hex;
 use proptest::prelude::*;
 use serde_json::Value;
